Move SmooveBikeRentalUpdater a Sandbox extention
