+++ docs/SandboxExtension.md
@@ -18,6 +18,7 @@ nearby stops generated by routing via OSM data.
 - [Transmodel API](sandbox/TransmodelApi.md) - Enturs GraphQL Transmodel API.
 - [SIRI updator](sandbox/SiriUpdator.md) - Update OTP with realtime information from a Transmodel SIRI data source.
 - [BikeRentalServiceDirectory](sandbox/BikeRentalServiceDirectory.md) - GBFS service directory endpoint.
+- [Smoove Bike Rental Updator Support](sandbox/SmooveBikeRental.md) - Smoove Bike Rental Updator(HSL)
 - [Mapbox Vector Tiles API](sandbox/MapboxVectorTilesApi.md) - Mapbox Vector Tiles API
 - [Park and Ride API](sandbox/ParkAndRideApi.md) - Park and Ride API
