+++ docs/sandbox/SmooveBikeRental.md
@@ -1,0 +1,13 @@
+# HSL Smoove Bike Rental Updater Support - OTP Sandbox Extension
+
+## Contact Info
+- Digitransit team, HSL, Helsinki, Finland
+
+## Changelog
+- Move this functonality into a sandbox
+
+## Documentation
+TODO
+
+### Configuration
+TODO
