+++ mkdocs.yml
@@ -46,6 +46,7 @@ pages:
     - Transmodel(NeTEx) GrpahQL API: 'sandbox/TransmodelApi.md'
     - SIRI Updater: 'sandbox/SiriUpdator.md'
     - Bike Rental Service Directory API support: 'sandbox/BikeRentalServiceDirectory.md'
+    - Smoove Bike Rental Updator Support: 'sandbox/SmooveBikeRental.md'
     - Mapbox Vector Tiles API: 'sandbox/MapboxVectorTilesApi.md'
     - Report API: 'sandbox/ReportApi.md'
     - Interactive OTP Launcher: 'sandbox/InteractiveOtpMain.md'
