+++ src/test/java/org/opentripplanner/updater/bike_rental/datasources/TestBikeRentalStationSource.java
@@ -36,51 +36,4 @@ public class TestBikeRentalStationSource extends TestCase {
         BikeRentalStation kergus = rentalStations.get(3);
         assertEquals("12", kergus.id);
     }
-
-    public void testSmoove() {
-        SmooveBikeRentalDataSource source = new SmooveBikeRentalDataSource(
-            new BikeRentalDataSourceParameters(
-                null,
-                "file:src/test/resources/bike/smoove.json",
-                null,
-                null,
-                Map.of()
-            )
-        );
-        assertTrue(source.update());
-        List<BikeRentalStation> rentalStations = source.getStations();
-
-        // Invalid station without coordinates shoulf be ignored, so only 3
-        assertEquals(3, rentalStations.size());
-        for (BikeRentalStation rentalStation : rentalStations) {
-            System.out.println(rentalStation);
-        }
-
-        BikeRentalStation hamn = rentalStations.get(0);
-        assertEquals("Hamn", hamn.name.toString());
-        assertEquals("A04", hamn.id);
-        // Ignore whitespace in coordinates string
-        assertEquals(24.952269, hamn.x);
-        assertEquals(60.167913, hamn.y);
-        assertEquals(11, hamn.spacesAvailable);
-        assertEquals(1, hamn.bikesAvailable);
-
-        BikeRentalStation fake = rentalStations.get(1);
-        assertEquals("Fake", fake.name.toString());
-        assertEquals("B05", fake.id);
-        assertEquals(24.0, fake.x);
-        assertEquals(60.0, fake.y);
-        // operative: false overrides available bikes and slots
-        assertEquals(0, fake.spacesAvailable);
-        assertEquals(0, fake.bikesAvailable);
-
-        BikeRentalStation foo = rentalStations.get(2);
-        assertEquals("Foo", foo.name.toString());
-        assertEquals("B06", foo.id);
-        assertEquals(25.0, foo.x);
-        assertEquals(61.0, foo.y);
-        assertEquals(5, foo.spacesAvailable);
-        assertEquals(5, foo.bikesAvailable);
-        // Ignores mismatch with total_slots
-    }
 }
